requirements.txt
deepgram-sdk>=3.2.3
sounddevice>=0.4.6
python-dotenv>=1.0.1
numpy>=1.26
requests>=2.31
openai>=1.37
webrtcvad>=2.0.10
pipecat-ai[silero,deepgram,openai,local]==0.0.46
loguru>=0.7.2
